feat: add individual demo pages
