+++ .gitignore
@@ -1,2 +1,3 @@
 dist
 demo
+node_modules
